refactor(runtime): share framework message renderers
