+++ CHANGELOG.md
@@ -65,6 +65,10 @@
 
 ### Performance Improvements
 
+- React and Solid now share one internal message renderer between their root
+  and `compiled` entrypoints. This removes parallel runtime implementations
+  while preserving the parser-free production boundary; lazy raw patterns use
+  a parse-only Core capability instead of re-entering catalog lookup.
 - Generated catalogs, transformed `Trans` components, and compiled MDX now use
   parser-free `compiled` package entrypoints. The real Vite MDX production
   proof drops from 214.65 kB to 209.25 kB raw JavaScript and from 68.28 kB to
