+++ packages/react/src/compiled-server.tsx
@@ -1,7 +1,7 @@
 import { getI18n } from "@palamedes/runtime"
 
-import { createCompiledTrans } from "./compiledShared"
+import { createTrans } from "./transShared"
 
-export { Fragment, type CompiledTransProps as TransProps } from "./compiledShared"
+export { Fragment, type TransProps } from "./transShared"
 
-export const Trans = createCompiledTrans(getI18n)
+export const Trans = createTrans(getI18n)
